# contracts/kiln-http-api-v1.openapi.json
{
  "openapi": "3.1.1",
...
  "info": {
    "title": "Kiln HTTP API",
    "version": "1.0.0",
    "description": "Canonical inventory of every production HTTP route, method, transport, handler owner, request type, response type, status, media type, path/header/query input, and structured error envelope. Every named payload component delegates its complete field contract to a canonical generated schema."
  },
  "servers": [
    {
      "url": "http://127.0.0.1:8420"
    }
  ],
...
  "paths": {
...
    "/v1/agent/runs": {
      "get": {
        "tags": [
          "agents"
        ],
        "summary": "List runs",
        "operationId": "agent_runs_list_runs_get_v1_agent_runs",
        "x-kiln-handler": "agent_runs::list_runs",
        "x-kiln-rust-signature": "async fn list_runs( State(state): State<AppState>, Query(q): Query<ListRunsQuery>, ) -> Result<Json<AgentRunListResponse>, ApiError>",
        "x-kiln-transport": "http",
        "x-kiln-query-rust-type": "ListRunsQuery",
        "responses": {
          "200": {
            "description": "Successful response.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AgentRunListResponse"
                }
              }
            },
            "x-kiln-rust-type": "AgentRunListResponse"
          },
          "default": {
            "$ref": "#/components/responses/ApiError"
          }
        }
      },
      "post": {
        "tags": [
          "agents"
        ],
        "summary": "Create run",
        "operationId": "agent_runs_create_run_post_v1_agent_runs",
        "x-kiln-handler": "agent_runs::create_run",
        "x-kiln-rust-signature": "async fn create_run( State(state): State<AppState>, Json(req): Json<CreateRunRequest>, ) -> Result<Json<AgentRunRecord>, ApiError>",
        "x-kiln-transport": "http",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateRunRequest"
              }
            }
          },
          "x-kiln-rust-type": "CreateRunRequest"
        },
        "responses": {
          "200": {
            "description": "Successful response.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AgentRunRecord"
                }
              }
            },
            "x-kiln-rust-type": "AgentRunRecord"
          },
          "default": {
            "$ref": "#/components/responses/ApiError"
          }
        }
      }
    },
    "/v1/agent/runs/status": {
      "get": {
        "tags": [
          "agents"
        ],
        "summary": "Runs status",
        "operationId": "agent_runs_runs_status_get_v1_agent_runs_status",
        "x-kiln-handler": "agent_runs::runs_status",
        "x-kiln-rust-signature": "async fn runs_status(State(state): State<AppState>) -> Json<AgentRunsStatusResponse>",
        "x-kiln-transport": "http",
        "responses": {
          "200": {
            "description": "Successful response.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AgentRunsStatusResponse"
                }
              }
            },
            "x-kiln-rust-type": "AgentRunsStatusResponse"
          },
          "default": {
            "$ref": "#/components/responses/ApiError"
          }
        }
      }
    },
    "/v1/agent/runs/{id}": {
      "get": {
        "tags": [
          "agents"
        ],
        "summary": "Get run",
        "operationId": "agent_runs_get_run_get_v1_agent_runs_id",
        "x-kiln-handler": "agent_runs::get_run",
        "x-kiln-rust-signature": "async fn get_run( State(state): State<AppState>, AxumPath(id): AxumPath<String>, ) -> Result<Json<AgentRunRecord>, ApiError>",
        "x-kiln-transport": "http",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "minLength": 1
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful response.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AgentRunRecord"
                }
              }
            },
            "x-kiln-rust-type": "AgentRunRecord"
          },
          "default": {
            "$ref": "#/components/responses/ApiError"
          }
        }
      }
    },
    "/v1/agent/runs/{id}/abort": {
      "post": {
        "tags": [
          "agents"
        ],
        "summary": "Abort run",
        "operationId": "agent_runs_abort_run_post_v1_agent_runs_id_abort",
        "x-kiln-handler": "agent_runs::abort_run",
        "x-kiln-rust-signature": "async fn abort_run( State(state): State<AppState>, AxumPath(id): AxumPath<String>, ) -> Result<Json<AgentRunAbortResponse>, ApiError>",
        "x-kiln-transport": "http",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "minLength": 1
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful response.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AgentRunAbortResponse"
                }
              }
            },
            "x-kiln-rust-type": "AgentRunAbortResponse"
          },
          "default": {
            "$ref": "#/components/responses/ApiError"
          }
        }
      }
    },
    "/v1/agent/runs/{id}/events": {
      "get": {
        "tags": [
          "agents"
        ],
        "summary": "Run events",
        "operationId": "agent_runs_run_events_get_v1_agent_runs_id_events",
        "x-kiln-handler": "agent_runs::run_events",
        "x-kiln-rust-signature": "async fn run_events( State(state): State<AppState>, AxumPath(id): AxumPath<String>, Query(q): Query<EventsQuery>, ) -> Result<Json<AgentRunEventsResponse>, ApiError>",
        "x-kiln-transport": "http",
        "x-kiln-query-rust-type": "EventsQuery",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "minLength": 1
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful response.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AgentRunEventsResponse"
                }
              }
            },
            "x-kiln-rust-type": "AgentRunEventsResponse"
          },
          "default": {
            "$ref": "#/components/responses/ApiError"
          }
        }
      }
    },
    "/v1/agent/runs/{id}/follow_up": {
      "post": {
        "tags": [
          "agents"
        ],
        "summary": "Follow up run",
        "operationId": "agent_runs_follow_up_run_post_v1_agent_runs_id_follow_up",
        "x-kiln-handler": "agent_runs::follow_up_run",
        "x-kiln-rust-signature": "async fn follow_up_run( State(state): State<AppState>, AxumPath(id): AxumPath<String>, Json(req): Json<MessageRequest>, ) -> Result<Json<AgentRunQueuedResponse>, ApiError>",
        "x-kiln-transport": "http",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "minLength": 1
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MessageRequest"
              }
            }
          },
          "x-kiln-rust-type": "MessageRequest"
        },
        "responses": {
          "200": {
            "description": "Successful response.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AgentRunQueuedResponse"
                }
              }
            },
            "x-kiln-rust-type": "AgentRunQueuedResponse"
          },
          "default": {
            "$ref": "#/components/responses/ApiError"
          }
        }
      }
    },
    "/v1/agent/runs/{id}/steer": {
      "post": {
        "tags": [
          "agents"
        ],
        "summary": "Steer run",
        "operationId": "agent_runs_steer_run_post_v1_agent_runs_id_steer",
        "x-kiln-handler": "agent_runs::steer_run",
        "x-kiln-rust-signature": "async fn steer_run( State(state): State<AppState>, AxumPath(id): AxumPath<String>, Json(req): Json<MessageRequest>, ) -> Result<Json<AgentRunQueuedResponse>, ApiError>",
        "x-kiln-transport": "http",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "minLength": 1
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MessageRequest"
              }
            }
          },
          "x-kiln-rust-type": "MessageRequest"
        },
        "responses": {
          "200": {
            "description": "Successful response.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AgentRunQueuedResponse"
                }
              }
            },
            "x-kiln-rust-type": "AgentRunQueuedResponse"
          },
          "default": {
            "$ref": "#/components/responses/ApiError"
          }
        }
      }
    },
...
  },
  "components": {
    "responses": {
      "ApiError": {
        "description": "Structured API failure.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ApiError"
            }
          }
        }
      }
    },
    "schemas": {
...
      "AgentRunAbortResponse": {
        "$ref": "kiln-control-plane-v1.schema.json#/$defs/AgentRunAbortResponse",
        "x-kiln-rust-type": "AgentRunAbortResponse",
        "x-kiln-field-schema-status": "complete"
      },
      "AgentRunEventsResponse": {
        "$ref": "kiln-control-plane-v1.schema.json#/$defs/AgentRunEventsResponse",
        "x-kiln-rust-type": "AgentRunEventsResponse",
        "x-kiln-field-schema-status": "complete"
      },
      "AgentRunListResponse": {
        "$ref": "kiln-control-plane-v1.schema.json#/$defs/AgentRunListResponse",
        "x-kiln-rust-type": "AgentRunListResponse",
        "x-kiln-field-schema-status": "complete"
      },
      "AgentRunQueuedResponse": {
        "$ref": "kiln-control-plane-v1.schema.json#/$defs/AgentRunQueuedResponse",
        "x-kiln-rust-type": "AgentRunQueuedResponse",
        "x-kiln-field-schema-status": "complete"
      },
      "AgentRunRecord": {
        "$ref": "kiln-control-plane-v1.schema.json#/$defs/AgentRunRecord",
        "x-kiln-rust-type": "AgentRunRecord",
        "x-kiln-field-schema-status": "complete"
      },
      "AgentRunsStatusResponse": {
        "$ref": "kiln-control-plane-v1.schema.json#/$defs/AgentRunsStatusResponse",
        "x-kiln-rust-type": "AgentRunsStatusResponse",
        "x-kiln-field-schema-status": "complete"
      },
...
      "ApiError": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "error"
        ],
        "properties": {
          "error": {
            "type": "object",
            "additionalProperties": false,
            "required": [
              "code",
              "message",
              "hint"
            ],
            "properties": {
              "code": {
                "type": "string"
              },
              "message": {
                "type": "string"
              },
              "hint": {
                "type": "string"
              }
            }
          }
        },
        "x-kiln-rust-type": "ApiError",
        "x-kiln-field-schema-status": "complete"
      },
...
      "CreateRunRequest": {
        "$ref": "kiln-control-plane-v1.schema.json#/$defs/CreateRunRequest",
        "x-kiln-rust-type": "CreateRunRequest",
        "x-kiln-field-schema-status": "complete"
      },
...
      "MessageRequest": {
        "$ref": "kiln-control-plane-v1.schema.json#/$defs/MessageRequest",
        "x-kiln-rust-type": "MessageRequest",
        "x-kiln-field-schema-status": "complete"
      },
...
    }
  },
...
}